refactor : 예외처리
