+++ src/main/java/com/backend/allreva/concert/exception/ConcertSearchNotFoundException.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.concert.exception;
+
+import com.backend.allreva.common.exception.CustomException;
+
+public class ConcertSearchNotFoundException extends CustomException {
+    public ConcertSearchNotFoundException() {
+        super(ConcertErrorCode.CONCERT_NOT_FOUND);
+    }
+}
