+++ src/main/java/com/backend/allreva/search/exception/ElasticSearchException.java
@@ -1,0 +1,9 @@
+package com.backend.allreva.search.exception;
+
+import com.backend.allreva.common.exception.CustomException;
+
+public class ElasticSearchException extends CustomException {
+    public ElasticSearchException() {
+        super(SearchErrorCode.ELASTICSEARCH_ERROR);
+    }
+}
